// src/app/interfaces/ivacante-detalle.ts
export interface IVacanteDetalle {
    idVacante: string;
    salario: string;
    imagen: string;
    estatus: string;
    nombre: string;
    destacado: boolean;
    descripcion: string;
    detalles: string;
    pais: string;
    nombreCategoria: string;
    descripcionCategoria: string;
    nombreEmpresa: string;
}